Change Porject-> mandays left to double

+++ model/src/main/java/com/media2359/euphoria/model/project/Project.java
@@ -36,7 +36,7 @@ public class Project implements java.io.Serializable{
 	private String name;
 	private String description;
 	private String projectManager;
-	private Integer manDaysLeft;
+	private Double manDaysLeft;
 	private Integer milestoneCount;
 	private Integer completedMilestoneCount;
 	private ProjectPlan projectPlan;
@@ -124,11 +124,11 @@ public class Project implements java.io.Serializable{
 	}
 	
 	@Column(name = "mandays_left")
-	public Integer getManDaysLeft() {
+	public Double getManDaysLeft() {
 		return manDaysLeft;
 	}
 
-	public void setManDaysLeft(Integer manDaysLeft) {
+	public void setManDaysLeft(Double manDaysLeft) {
 		this.manDaysLeft = manDaysLeft;
 	}
